Converted enable_single_byte_value_storage to be a template variable.
Replaced the global BOOST_OUTCOME_ALLOW_THROWING_MOVE_TYPES macro disable with per type disable via enable_move_throwing_type<T>.

Fix an ambiguous resolution of make_expected_from_error.

+++ test/expected_pass.cpp
@@ -600,7 +600,7 @@ void expected_from_error()
 void expected_from_error_U()
 {
   // From stde::unexpected_type constructor.
-  auto e = stde::make_expected_from_error<std::string, short>(42);
+  auto e = stde::make_expected_from_error<std::string, short>((short) 42);
   static_assert(std::is_same<decltype(e), stde::expected<std::string, short>>{}, "");
   BOOST_CHECK_EQ(static_cast<bool>(e), false);
 }
